requirements.txt
psycopg2
Mastodon.py
